+++ src/routes/Details/Data/detailsData.tsx
