+++ src/routes/Details/Styles/DetailsStyle.css
