+++ src/routes/HobbiesPage/Hobbies.tsx
@@ -13,58 +13,20 @@ import {
   TextInput,
   Textarea,
 } from "@mantine/core";
-// import { hobbies as initialHobbies } from "./data"; // Import existing hobbies
+
 import "./HobbiesStyle.css";
 import { QueryClient, useQueries, useQuery } from "@tanstack/react-query";
-import {getTestHobby } from "./queries/query";
+import { getTestHobby } from "./queries/query";
 
 const Hobbies = ({ setMain }: { setMain: (main: string) => void }) => {
-  // State to manage hobbies
-
-  const { data: hobbies, error } = useQuery(
-    {
+  
+  const { data: hobbies, error } = useQuery({
     queryKey: ["testHobby"],
     queryFn: getTestHobby,
-  },
-);
-
-
-  // Hardcoded new hobby
-  const newHobby = {
-    title: "Photography",
-    content:
-      "Photography allows me to capture the world around me, freezing moments in time and creating memories that last forever.",
-    image: "/photography.jpg",
-  };
-
-  // Add hardcoded new hobby to the list
-  // const addHobby = () => {
-  //   setHobbies([...hobbies, newHobby]);
-  // };
+  });
 
   return (
     <div>
-      {/* Button to add the hardcoded hobby */}
-      {/* <Button onClick={addHobby} style={{ marginBottom: "20px" }}>
-        Add Photography Hobby
-      </Button> */}
-      {/* {testHobby?.map((hobby,key) => {
-        return (
-          <>
-          <Text> {hobby.id}</Text>
-          <Text> {hobby.name}</Text>
-          <Text> {hobby.surname}</Text>
-           </>
-        )
-
-      })} */}
-      {/* <Button
-        style={{ marginBottom: "20px" }}
-        onClick={() => console.log(testHobby)}
-      >
-        Get all users
-      </Button> */}
-
       <Card className="hobbies-card">
         {hobbies?.map((hobby, key) => {
           const isEven = Boolean(!(hobbies.indexOf(hobby) % 2));
