+++ src/routes/HobbiesPage/Models/models.tsx
@@ -1,9 +1,3 @@
-export type IHobby = {
-  id: number;
-  name: string;
-  surname: string;
-};
-
 export type Hobby = {
     id: number;
     title:string;
